move utilities.js to a new folder
